server/pyproject.toml
[project]
name = "research-assistant"
version = "0.1.0"
description = "Voice research assistant demo powered by Pipecat Subagents"
requires-python = ">=3.11"
dependencies = [
    "pipecat-ai[webrtc,daily,silero,soniox,cartesia,openai,runner]>=0.0.108",
    "pipecat-ai-subagents[examples]>=0.0.1",
    "claude-agent-sdk>=0.1.0",
    "python-dotenv>=1.0.1",
]

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["I"]


[tool.uv.sources]
pipecat-ai-subagents = { path = "../../pipecat-subagents", editable = true }
pipecat-ai = { path = "../../pipecat", editable = true }